requirements.txt
zappa
flask
request_oauthlib
python-twitter
boto3
flask-testing
mock